+++ tests/test_content_generator.cpp
@@ -5,6 +5,7 @@
 #include <sstream>
 
 #include "cli/content_generator.hpp"
+#include "test_util.hpp"
 
 namespace fs = std::filesystem;
 
@@ -20,8 +21,7 @@ struct ContentFixture {
 
     ContentFixture() {
         saved_cwd = fs::current_path().string();
-        root_dir = (fs::temp_directory_path() /
-                   ("cstatic_cg_" + std::to_string(std::rand()))).string();
+        root_dir = cstatic_test::unique_temp_dir("cstatic_cg_");
         fs::create_directories(root_dir + "/src");
         fs::create_directories(root_dir + "/archetypes");
         fs::current_path(root_dir);