+++ tests/test_shortcodes.cpp
@@ -5,6 +5,7 @@
 
 #include "content/shortcodes.hpp"
 #include "utils/file_io.hpp"
+#include "test_util.hpp"
 
 namespace fs = std::filesystem;
 
@@ -20,8 +21,7 @@ struct ShortcodeFixture {
 
     ShortcodeFixture() {
         saved_cwd = fs::current_path().string();
-        root_dir = (fs::temp_directory_path() /
-                   ("cstatic_sc_" + std::to_string(std::rand()))).string();
+        root_dir = cstatic_test::unique_temp_dir("cstatic_sc_");
         fs::create_directories(root_dir);
         fs::create_directories(root_dir + "/shortcodes");
         fs::current_path(root_dir);